backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
psycopg2-binary==2.9.6
geoalchemy2==0.13.0
python-dotenv==1.0.0
flask-jwt-extended==4.4.4